+++ biopytools/kmertools/kmer_count/jellyfish_processor.py
@@ -6,7 +6,6 @@ import os
 import subprocess
 from pathlib import Path
 from typing import List
-from modelscope.hub.errors import FileIntegrityError
 
 class JellyfishProcessor:
     """Jellyfish处理器|Jellyfish Processor"""